pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "AI Resume Generator with LLaMA integration"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "weasyprint>=66.0",
    "requests>=2.31.0",
]